Adapted for Squid and JJ Simulations
In version 003 I added a squid loop to make it so that there could be a
jump in the phase of the second half of the junction.  This worked well
but was clunky.  I then adapted version 004 so that it could cleanly
handle a squid loop and two junctions of different areas and critical
currents.  These simulations still assume a small BetaL parameter, adn
as a result do not include any screening currents contributing to the
phase in the squid loop.