Comparison in Plot with Jam variable
I added an extra loop on the outside of the meat of the program so that
it would run once with Jam=0, and then once with jam=1.  That made it so
that when it plots its easy to compare the conventional plot to any
plots with added CPR or noise features.