Added Alpha for Plotting Ic vs Phi for different parameters
Added more complexity to the j loop to make it so that you can run
through an arbitrary number of steps (before it was just 1 or 0) of the
parameter alpha (it was Jam before).  It makes it really easy to plot
all of the Ic vs Phi for a range of values for a parameter.  Its
currently set up to test the junction asymmetry in sin2Phi junctions