Added Asymetrical Squid Programs in V005 and V006
Two iterations of the Asymetrical Squid are added.  One uses for loops,
and the other enables the use of parfor because the phase loop was
included in the arrays rather than as a separate for loop.  V006 also
includes more information because it plots the phase as well as the Ic