Adding Dwave Phase Shift
In the CPR I added a parameter that could be tuned by the Alpha and made
two changes to try and simulate some data we had from a year ago.
Asymetrical squid with sin 2Phi CPR in the small junction, a large
second junction, and a squid loop that is 100 times bigger than the
junction areas.  It looks alot like the data, and I set it up to run
through alpha parameters to see what the asymmetry of the junction
needes to be to get the junction behavior we saw.