Added v000 with just the Phase 1 Loop
This version is to show that with different Phase1 there is more or less
supercurrent across the junction.  If you change the flux in the
junction the pattern shifts, although the overall shape is the same.  In
later versions the program just finds the maximum and records that
rather than recording the whole SCurrentNet vs Phase1 vector.