Added Plotting of Phase at Max and Min
The addition was added to the squid simulation with a for loop to help
diagnose why we were not seing the mapping of the CPR with a sin phi/w
term.  The junction that Phase1 was set at was switched to 2 so that the
Phases made more sense when looking at them.